backend/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
asyncpg==0.27.0
pydantic==2.5.2
pydantic-settings==0.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
alembic==1.11.1
